refactor: redisJobOptions as type
