+++ src/classes/compat.ts
@@ -27,7 +27,6 @@ import { Queue } from './queue';
 import { Worker } from './worker';
 import { QueueEvents } from './queue-events';
 import {
-  JobsOptions,
   QueueOptions,
   RepeatOptions,
   QueueEventsOptions,
@@ -35,6 +34,7 @@ import {
   WorkerOptions,
   Processor,
 } from '../interfaces';
+import { JobsOptions } from '../types';
 
 type CommonOptions = QueueSchedulerOptions &
   QueueOptions &
