+++ src/interfaces/parent.ts
@@ -1,4 +1,4 @@
-import { JobsOptions } from './jobs-options';
+import { JobsOptions } from '../types';
 
 /**
  * Describes the parent for a Job.
