+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './finished-target';
 export * from './job-json-sandbox';
+export * from './job-options';
 export * from './job-type';
